LibGfx: add constexpr to constructors for Point and Rect
This is the only change so far made to any files outside of Roguelike.

It adds constexpr to constructors for Point and Rect, which Roguelike
uses in several places. In all likelihood, these changes will be
reverted at a later date, as it was purely an experiment, and Roguelike
should work fine without them (although it will require removing
constexpr from several functions within Roguelike).